fix: multiple bug fixes and code refactoring (#27)
* fix: correct euclidean distance and inner product computation logic for sq4 in dist_ip.hpp and dist_l2.hpp

fix: fix bitmap allocation size bug in sequential_storage.hpp

refactor: make mem_prefetch_l1 and mem_prefetch_l2 const-correct in prefetch.hpp

refactor: use size_t for alignment function do_align() in types.hpp

refactor: make data loader template-aware with sizeof(T)

fix: correct Neighbor template parameter order to IDType, DistanceType in query_utils.hpp

*refactor: delete redundent code

---------

+++ include/utils/types.hpp
@@ -16,13 +16,14 @@
 
 #pragma once
 
+#include <cstddef>
 #include <cstdint>
 #include <stdexcept>
 
 using WorkerID = uint32_t;
 using CpuID = uint32_t;
 namespace alaya {
-inline auto do_align(uint32_t size, uint32_t align_num) -> uint32_t {
+inline auto do_align(size_t size, size_t align_num) -> size_t {
   // Ensure align_num is not zero (to avoid division by zero)
   if (align_num == 0) {
     throw std::invalid_argument("align_num cannot be zero");
